refactor(common-utility): Moves file sytem name helpers implementations to their OS-specific C++ files (Clément's review comment)
