+++ src/libcommon/utility/utility.cpp
@@ -220,8 +220,8 @@ const std::string &CommonUtility::versionTag() {
 uint64_t CommonUtility::versionBuild() {
     return KDRIVE_VERSION_BUILD;
 }
-namespace {
-std::string getRootFsType(const SyncPath &targetPath) {
+
+std::string CommonUtility::getRootFsType(const SyncPath &targetPath) {
     static std::unordered_map<std::string, std::string> rootFsTypeMap;
 
     auto it = rootFsTypeMap.find(targetPath.root_name().string());
@@ -234,7 +234,7 @@ std::string getRootFsType(const SyncPath &targetPath) {
     }
     return it->second;
 }
-} // namespace
+
 
 bool CommonUtility::isNTFS(const SyncPath &targetPath) {
     static const std::string ntfs("NTFS");
@@ -274,94 +274,6 @@ bool CommonUtility::isLiteSyncCompatible([[maybe_unused]] const SyncPath &target
 #else
     return false;
 #endif
-}
-
-#if defined(KD_LINUX)
-namespace {
-#ifndef EXFAT_SUPER_MAGIC
-#define EXFAT_SUPER_MAGIC 0x2011BAB0
-#endif
-
-constexpr auto exFat = "exFAT";
-constexpr auto ext234 = "EXT2/3/4";
-
-std::string formatFsName(const std::string &prettyName, const int64_t fType) {
-    std::stringstream stream;
-    stream << std::hex << fType;
-    return prettyName + " | 0x" + stream.str();
-}
-} // namespace
-
-bool CommonUtility::isEXT234(const SyncPath &targetPath) {
-    return contains(getRootFsType(targetPath), ext234);
-}
-
-std::string CommonUtility::exFAT() {
-    return formatFsName(exFat, EXFAT_SUPER_MAGIC);
-}
-
-#endif
-
-std::string CommonUtility::fileSystemName(const SyncPath &targetPath) {
-#if defined(KD_MACOS)
-    struct statfs stat;
-    if (statfs(targetPath.root_path().native().c_str(), &stat) == 0) {
-        return stat.f_fstypename;
-    }
-#elif defined(KD_WINDOWS)
-    TCHAR szFileSystemName[MAX_PATH + 1];
-    DWORD dwMaxFileNameLength = 0;
-    DWORD dwFileSystemFlags = 0;
-
-    if (GetVolumeInformation(targetPath.root_path().c_str(), NULL, 0, NULL, &dwMaxFileNameLength, &dwFileSystemFlags,
-                             szFileSystemName, sizeof(szFileSystemName)) == TRUE) {
-        return ws2s(szFileSystemName);
-    } else {
-        // Not all the requested information is retrieved
-        DWORD dwError = GetLastError();
-        std::wstringstream message;
-        message << L"Error in GetVolumeInformation for " << Path2WStr(targetPath.root_name()) << L" ("
-                << utility_base::getErrorMessage(dwError) << L")";
-        sentry::Handler::captureMessage(sentry::Level::Warning, "CommonUtility::fileSystemName", ws2s(message.str()));
-
-        // !!! File system name can be OK or not !!!
-        return ws2s(szFileSystemName);
-    }
-#elif defined(KD_LINUX)
-    struct statfs stat;
-    if (statfs(targetPath.root_path().native().c_str(), &stat) == 0) {
-        switch (stat.f_type) {
-            case EXFAT_SUPER_MAGIC:
-                return exFAT();
-            case 0x137d:
-                return formatFsName("EXT(1)", stat.f_type);
-            case 0xef51:
-                return formatFsName("EXT2", stat.f_type);
-            case 0xef53: // EXT_SUPER_MAGIC
-                return formatFsName(ext234, stat.f_type);
-            case 0xbad1dea:
-            case 0xa501fcf5:
-            case 0x58465342:
-                return formatFsName("XFS", stat.f_type);
-            case 0x9123683e:
-            case 0x73727279:
-                return formatFsName("BTRFS", stat.f_type);
-            case 0xf15f:
-                return formatFsName("ECRYPTFS", stat.f_type);
-            case 0x4244:
-                return formatFsName("HFS", stat.f_type);
-            case 0x5346544e:
-                return formatFsName("NTFS", stat.f_type);
-            case 0x858458f6:
-                return formatFsName("RAMFS", stat.f_type);
-            default:
-                return formatFsName("Unknown-see corresponding entry at https://man7.org/linux/man-pages/man2/statfs.2.html",
-                                    stat.f_type);
-        }
-    }
-#endif
-
-    return "UNIDENTIFIED";
 }
 
 void CommonUtility::resetTranslations() {
