+++ src/libcommon/utility/utility_linux.cpp
@@ -18,21 +18,14 @@
 
 #include "utility.h"
 
-#include <QStandardPaths>
-
 #include <filesystem>
 #include <fstream>
 #include <unistd.h>
 #include <sys/types.h>
+#include <sys/vfs.h>
 #include <pwd.h>
 
 #include "utility/types.h"
-
-#include <log4cplus/logger.h>
-#include <log4cplus/loggingmacros.h>
-
-#include <Poco/File.h>
-#include <Poco/Exception.h>
 
 namespace KDC {
 
@@ -114,6 +107,67 @@ std::string CommonUtility::osVersion() {
 
 std::string CommonUtility::distributionName() {
     return extractOSInfo("NAME");
+}
+
+
+namespace {
+#ifndef EXFAT_SUPER_MAGIC
+#define EXFAT_SUPER_MAGIC 0x2011BAB0
+#endif
+
+constexpr auto exFat = "exFAT";
+constexpr auto ext234 = "EXT2/3/4";
+
+std::string formatFsName(const std::string &prettyName, const int64_t fType) {
+    std::stringstream stream;
+    stream << std::hex << fType;
+    return prettyName + " | 0x" + stream.str();
+}
+} // namespace
+
+bool CommonUtility::isEXT234(const SyncPath &targetPath) {
+    return contains(getRootFsType(targetPath), ext234);
+}
+
+std::string CommonUtility::exFAT() {
+    return formatFsName(exFat, EXFAT_SUPER_MAGIC);
+}
+
+std::string CommonUtility::fileSystemName(const SyncPath &targetPath) {
+    struct statfs stat;
+
+    if (statfs(targetPath.root_path().native().c_str(), &stat) == 0) {
+        switch (stat.f_type) {
+            case EXFAT_SUPER_MAGIC:
+                return exFAT();
+            case 0x137d:
+                return formatFsName("EXT(1)", stat.f_type);
+            case 0xef51:
+                return formatFsName("EXT2", stat.f_type);
+            case 0xef53: // EXT_SUPER_MAGIC
+                return formatFsName(ext234, stat.f_type);
+            case 0xbad1dea:
+            case 0xa501fcf5:
+            case 0x58465342:
+                return formatFsName("XFS", stat.f_type);
+            case 0x9123683e:
+            case 0x73727279:
+                return formatFsName("BTRFS", stat.f_type);
+            case 0xf15f:
+                return formatFsName("ECRYPTFS", stat.f_type);
+            case 0x4244:
+                return formatFsName("HFS", stat.f_type);
+            case 0x5346544e:
+                return formatFsName("NTFS", stat.f_type);
+            case 0x858458f6:
+                return formatFsName("RAMFS", stat.f_type);
+            default:
+                return formatFsName("Unknown-see corresponding entry at https://man7.org/linux/man-pages/man2/statfs.2.html",
+                                    stat.f_type);
+        }
+    }
+
+    return "UNIDENTIFIED";
 }
 
 } // namespace KDC
